Move TorTestHelper to package helpers
